fix: move network status to separate folder
