+++ packages/sync/src/offline/index.ts
@@ -1,5 +1,5 @@
-export * from "./NetworkStatus";
-export * from "./WebNetworkStatus";
-export * from "./CordovaNetworkStatus";
+export * from "./network/NetworkStatus";
+export * from "./network/WebNetworkStatus";
+export * from "./network/CordovaNetworkStatus";
 export * from "./OfflineQueueListener";
 export * from "./OfflineRestoreHandler";
