+++ packages/sync/src/offline/network/CordovaNetworkStatus.ts
