+++ packages/sync/src/offline/network/NetworkStatus.ts
