+++ packages/sync/src/offline/network/WebNetworkStatus.ts
